+++ src/main/java/domain/movements/Direction.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain.movements;
 
 public enum Direction {
     NORTH(1, 0),
