+++ src/test/java/domain/BoardTest.java
@@ -1,9 +1,10 @@
 package domain;
 
-import domain.piece.Piece;
-import java.util.Map;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
+
+import domain.pieces.Piece;
+import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
