+++ src/test/java/domain/piece/CannonTest.java
@@ -2,8 +2,13 @@ package domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.PieceOnRoute;
 import domain.Point;
 import domain.Team;
+import domain.pieces.Cannon;
+import domain.pieces.Chariot;
+import domain.pieces.EmptyPiece;
+import domain.pieces.Piece;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
