+++ src/test/java/domain/piece/RouteTest.java
@@ -1,6 +1,8 @@
 package domain.piece;
 
 import domain.Point;
+import domain.movements.Direction;
+import domain.movements.Route;
 import java.util.ArrayList;
 import java.util.List;
 import org.assertj.core.api.Assertions;
