+++ src/test/java/domain/piece/SoldierTest.java
@@ -1,9 +1,14 @@
 package domain.piece;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import domain.PieceOnRoute;
 import domain.Point;
 import domain.Team;
+import domain.pieces.EmptyPiece;
+import domain.pieces.Piece;
+import domain.pieces.Soldier;
 import java.util.List;
-import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
